implemented methods to create user, check account balance, and credit…

+++ src/main/java/com/demo/finbank/dto/requests/CreditDebitRequest.java
@@ -13,5 +13,6 @@ import java.math.BigDecimal;
 @Builder
 public class CreditDebitRequest {
     private String accountNumber;
-    private BigDecimal amount;
+    //private BigDecimal amount;
+    private int amount;
 }
